+++ packages/terra-list/src/ListItem.jsx
@@ -54,6 +54,10 @@ const propTypes = {
    * @private Callback function not intended for use with this API, but if set pass it through to the element regardless.
    */
   onMouseDown: PropTypes.func,
+  /**
+   * Set to true for non-selectable items in a list having role as listbox.
+   */
+  isRoleListbox: PropTypes.bool,
 };
 
 const defaultProps = {
@@ -61,6 +65,7 @@ const defaultProps = {
   hasChevron: false,
   isSelected: false,
   isSelectable: false,
+  isRoleListbox: false,
 };
 
 const ListItem = ({
@@ -75,6 +80,7 @@ const ListItem = ({
   onMouseDown,
   onSelect,
   refCallback,
+  isRoleListbox,
   ...customProps
 }) => {
   const listItemClassNames = cx([
@@ -94,7 +100,7 @@ const ListItem = ({
     attrSpread['data-item-show-focus'] = 'true';
     attrSpread.onBlur = ListUtils.wrappedEventCallback(onBlur, event => event.currentTarget.setAttribute('data-item-show-focus', 'true'));
     attrSpread.onMouseDown = ListUtils.wrappedEventCallback(onMouseDown, event => event.currentTarget.setAttribute('data-item-show-focus', 'false'));
-  } else if (customProps.listrole === 'listbox') {
+  } else if (isRoleListbox) {
     attrSpread.role = 'option';
     attrSpread['aria-selected'] = isSelected;
     attrSpread['aria-disabled'] = 'true';